updating before I mess up some more

+++ DL/Entities/LineItem.cs
@@ -8,7 +8,7 @@ namespace DL.Entities
     public partial class LineItem
     {
         public int Id { get; set; }
-        public decimal? Quantity { get; set; }
+        public int? Quantity { get; set; }
         public int? ProductId { get; set; }
         public int? OrderId { get; set; }
